Improve timeseries data parsing
Timeseries parsing is redefined in a more automated way by:
- restructuring and renaming the input files according to given rules
- adding a `parse_timeseries` function

Specifically, the input data folder must be structured in the following way:
- input timeseries files must be organized in folders named after `HVDCWISE_TEA` network component name of reference
- input .csv files must be named using the same `HVDCWISE_TEA` internal variable name the user wants to modify

Each file is structured as a matrix of size (T+1)xN, where T is the numer of timestamps and N the number of components
for which a timeseries is specified. The first row of the file specified the component index.

An example of folder and file structure is provided.